fandango-tests/pyproject.toml
[project]
name = "fandango-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fandango-fuzzer>=0.8.1",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
